Update Naming
MachineDetectionRequest -> MachineDetectionConfiguration
