+++ bandwidth/voice/models/create_call_request.py
@@ -6,7 +6,7 @@ bandwidth
 This file was automatically generated by APIMATIC v3.0 (
  https://www.apimatic.io ).
 """
-from bandwidth.voice.models.machine_detection_request import MachineDetectionRequest
+from bandwidth.voice.models.machine_detection_configuration import MachineDetectionConfiguration
 
 
 class CreateCallRequest(object):
@@ -40,7 +40,7 @@ class CreateCallRequest(object):
             here.
         tag (string): TODO: type description here.
         application_id (string): TODO: type description here.
-        machine_detection (MachineDetectionRequest): TODO: type description
+        machine_detection (MachineDetectionConfiguration): TODO: type description
             here.
 
     """
@@ -143,7 +143,7 @@ class CreateCallRequest(object):
         disconnect_url = dictionary.get('disconnectUrl')
         disconnect_method = dictionary.get('disconnectMethod')
         tag = dictionary.get('tag')
-        machine_detection = MachineDetectionRequest.from_dictionary(dictionary.get('machineDetection')) if dictionary.get('machineDetection') else None
+        machine_detection = MachineDetectionConfiguration.from_dictionary(dictionary.get('machineDetection')) if dictionary.get('machineDetection') else None
 
         # Return an object of this model
         return cls(mfrom,
